refactor: component is job specific resource
