+++ simaple/data/jobs/resources/components/adele.yaml
