+++ simaple/data/jobs/resources/components/adventurer.yaml
