+++ simaple/data/jobs/resources/components/archmagefb.yaml
