+++ simaple/data/jobs/resources/components/archmagetc.yaml
