+++ simaple/data/jobs/resources/components/bishop.yaml
