+++ simaple/data/jobs/resources/components/classes/pirate.yaml
