+++ simaple/data/jobs/resources/components/common.yaml
