+++ simaple/data/jobs/resources/components/common_v.yaml
