+++ simaple/data/jobs/resources/components/cygnus.yaml
