+++ simaple/data/jobs/resources/components/dualblade.yaml
