+++ simaple/data/jobs/resources/components/magician.yaml
