+++ simaple/data/jobs/resources/components/mechanic.yaml
