+++ simaple/data/jobs/resources/components/mob.yaml
