+++ simaple/data/jobs/resources/components/resistance.yaml
