+++ simaple/data/jobs/resources/components/seedring.yaml
