+++ simaple/data/jobs/resources/components/soulmaster.yaml
