+++ simaple/data/jobs/resources/components/thief.yaml
